% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine output of the swapping algorithm}
\source{
Puccetti, G., Rüschendorf, L. & Vanduffel, S. (2020). On the
  computation of Wasserstein barycenters, Journal of Multivariate Analysis,
  176.
}
\usage{
combine(x)
}
\arguments{
\item{x}{a three dimensional array (rows = subsets, columns = par, slices
= samples) containing posterior samples for all subsets}
}
\value{
A \code{wasp} object, which can be further analyzed using the
  associated function \code{\link{summary.wasp}}.
}
\description{
This (non-exported) function combines the output from the swapping algorithm (Puccetti,
Rüschendorf and Vanduffel, 2020).
}
