% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hpd_est}
\alias{hpd_est}
\title{Compute the 95 percent Highest Posterior Density interval}
\source{
Venter, J.H. (1967). On estimation of the mode, Annals of
  Mathematical Statistics, 38(5), 1446-1455.
}
\usage{
hpd_est(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A vector containing the lower and upper bound of the 96% Highest
  Posterior Density interval of a numeric vector as computed by the methods
  from Venter (1967).
}
\description{
Compute the 95 percent Highest Posterior Density interval
}
\examples{
library(waspr)
hpd_est(pois_logistic[1,1,])

}
