% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode_est}
\alias{mode_est}
\title{Compute the mode}
\source{
Venter, J.H. (1967). On estimation of the mode, Annals of
  Mathematical Statistics, 38(5), 1446-1455.
}
\usage{
mode_est(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
The mode of a numeric vector as computed by the methods from Venter
  (1967).
}
\description{
Compute the mode
}
\examples{
library(waspr)
mode_est(pois_logistic[1,1,])

}
