% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pois_logistic}
\alias{pois_logistic}
\title{pois_logistic}
\format{
An array with 3 dimensions of which the first represents the
 subposteriors (size = 8), the second represents the paramters (size = 8) and
 the third represents the amount of mcmc samples (size = 450).
}
\usage{
pois_logistic
}
\description{
A set of mcmc samples from 8 subposteriors from the analysis of a joint model
with a logistic and poisson outcome variable.
}
\keyword{datasets}
