% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{summary}
\alias{summary}
\title{Posterior summaries for the Wasserstein barycenter of subset posteriors}
\usage{
summary(object)
}
\arguments{
\item{object}{a wasp object.}
}
\description{
\code{summary} gives a posterior summary (mean, mode, sd, HPD)
}
\details{
the method \link[waspr]{summary.wasp} has its own help page.
}
\examples{
library(waspr)

}
