% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waspr.R
\docType{package}
\name{waspr}
\alias{waspr}
\title{waspr: an R package for computing Wasserstein barycenters of subset
posteriors}
\source{
Puccetti, G., Rüschendorf, L. & Vanduffel, S. (2020). On the
  computation of Wasserstein barycenters, Journal of Multivariate Analysis,
  176.
}
\description{
This package contains functions to compute Wasserstein barycenters of subset
posteriors using the swapping algorithm developed by Puccetti, Rüschendorf
and Vanduffel (2020).
}
\section{Functions}{


  The main functions of the package are:

  \code{\link{wasp}}, which runs the swapping algorithm developed by
  Puccetti, Rüschendorf and Vanduffel (2020), combines the output from the
  swapping algorithm and computes the Wasserstein barycenter. It returns an
  S3 object of type \code{wasp}, which can be further analyzed through
  associated functions.

  \code{\link{combine}}, which
}

