% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{summary}
\alias{summary}
\title{Posterior summaries for the Wasserstein barycenter of subset posteriors}
\usage{
summary(x)
}
\arguments{
\item{x}{a \code{wasp} object.}
}
\description{
\code{summary} gives a posterior summary (mean, mode, sd, HPD)
}
\details{
the method \link[waspr]{summary.wasp} has its own help page.
}
\examples{
library(waspr)

}
